+++ .gitignore
@@ -22,4 +22,4 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
-server/data/*.json
+/server/data/*.json
