+++ server/data/bag.json
@@ -1,1 +1,1 @@
-[]
+[{"brandName":"MARCELO BURLON COUNTY","id":"1","brandId":36,"subtitle":"Logo appliquè sweatshirt","price":"210","priceDiscounted":"137.50","image":"images/img01.jpg","sortPrice":"137.50"},{"brandName":"MARCELO BURLON COUNTY","id":"51","brandId":44,"subtitle":"Logo appliquè sweatshirt","price":"210","priceDiscounted":"137.50","image":"images/img01.jpg","sortPrice":"137.50"},{"brandName":"NEIL BARRETT","id":"22","brandId":35,"subtitle":"Spray paint checked shirt","price":"335","image":"images/img01.jpg","sortPrice":"335"},{"brandName":"MARCELO BURLON COUNTY","id":"57","brandId":54,"subtitle":"Spray paint checked shirt","price":"335","image":"images/img01.jpg","sortPrice":"335"},{"brandName":"NEIL BARRETT","id":"12","brandId":55,"subtitle":"Spray paint checked shirt","price":"335","image":"images/img01.jpg","sortPrice":"335"},{"brandName":"NEIL BARRETT","id":"31","brandId":25,"subtitle":"Logo appliquè sweatshirt","price":"210","priceDiscounted":"137.50","image":"images/img01.jpg","sortPrice":"137.50"},{"brandName":"OFF-WHITE","id":"9","brandId":38,"subtitle":"Lightning bolt print T-shirt","price":"175","image":"images/img01.jpg","sortPrice":"175"}]
