+++ server/modules/cart/router.js
@@ -8,65 +8,16 @@ const cartController = require("./controller");
 /**
  * Route to Get The Shopping Cart
  */
-router.get("/cart", async (req, res) => {
-
-	const requestConfiguration = {
-		responseType: "JSON"
-	}
-	let subtotal = 0;
-
-	const shoppingCart = await utils.readFile("./server/data/bag.json", requestConfiguration);
-
-	 shoppingCart.forEach(product => {
-		if(product.hasOwnProperty("priceDiscounted")) {
-			subtotal += Number(product.priceDiscounted);
-			
-		} else {
-			subtotal += Number(product.price);
-		}
-			 
-	});
-
-	return res.status(200).send({products:shoppingCart,subtotal: subtotal});
-});
+router.get("/cart", cartController.getCart);
 
 /**
  * Route to Add items to the Cart
  */
-router.post("/cart", async (req, res) => {
-	const requestConfiguration = {
-		responseType: "JSON"
-	}
-
-	const product = req.body.product;
-
-	let shoppingCart = utils.readFile("./server/data/bag.json", requestConfiguration);
-	
-	shoppingCart = [...shoppingCart, product];
-
-	const response = await utils.writeFile("./server/data/bag.json", JSON.stringify(shoppingCart), {})
-
-	return res.status(200).send(response);
-
-});
+router.post("/cart", cartController.saveOnCart);
 
 /**
  * Route to Remove items to the Cart
  */
-router.delete("/cart", async (req, res) => {
-
-	const requestConfiguration = {
-		responseType: "JSON"
-	}
-
-	const productId = req.body.productId;
-
-	let shoppingCart = utils.readFile("./server/data/bag.json", requestConfiguration);
-	const newShopingCart = shoppingCart.filter(product => Number(product.id) !== Number(productId));
-	
-	const response = await utils.writeFile("./server/data/bag.json", JSON.stringify(newShopingCart), {})
-	return res.status(200).send(response);
-
-})
+router.delete("/cart", cartController.removeFromCart)
 
 module.exports = router;
