+++ server/router/router.js
@@ -1,12 +1,9 @@
 const express = require("express");
 const router = express.Router();
 
-const utils = require("./../helpers/utils");
-
 const productsRouter = require("./../modules/products/router");
 const brandsRouter = require("./../modules/brands/router");
 const cartRouter = require("./../modules/cart/router");
-
 
 /**
  * Routes
@@ -14,7 +11,6 @@ const cartRouter = require("./../modules/cart/router");
 router.get("/", (req, res) => {
 	res.status(200).send("The webserver is ON");
 });
-
 
 router.use("/", productsRouter);
 router.use("/", brandsRouter);
